fix(scanner): address PR0 review — fix README regen command + decouple _normalize (#225 PR0)
- README pointed to wrong pytest node ID; corrected to test_capture_btcusdt.
- _normalize moved from capture_baseline.py to scanner_frozen.py so the
  snapshot test no longer transitively imports the one-shot capture script.
